--- Joker/src/themes/index.ts
@@ -1,2 +1,0 @@
-export * from './theme-interface';
-export * from './themes';
