--- Joker/src/themes/theme-interface.ts
@@ -1,36 +1,0 @@
-export interface ColorTheme {
-  transparent: string;
-  textBase1: string;
-  textBase2: string;
-  textBase3: string;
-  textSecond1: string;
-  textSecond2: string;
-  textGreen1: string;
-  textRed1: string;
-  textYellow1: string;
-  textBlue1: string;
-  bg1: string;
-  bg2: string;
-  bg3: string;
-  bg4: string;
-  bg5: string;
-  bg6: string;
-  bg7: string;
-  bg8: string;
-  bg9: string;
-  bg10: string;
-  graphicBase1: string;
-  graphicBase2: string;
-  graphicBase3: string;
-  graphicRed1: string;
-  graphicGreen1: string;
-  graphicGreen2: string;
-  graphicSecond1: string;
-  graphicSecond2: string;
-  graphicSecond3: string;
-  graphicSecond4: string;
-}
-export interface Theme {
-  id: string;
-  color: ColorTheme;
-}
